implement stdarg.h

+++ include/stddef.h
@@ -5,8 +5,8 @@
 
 #define NULL ((void *)0)
 
-typedef unsigned long size_t;
-typedef long ptrdiff_t;
+typedef unsigned int size_t;
+typedef int ptrdiff_t;
 typedef unsigned int wchar_t;
 typedef long double max_align_t;
